refactor(practice): move mergePracticeNotesByTick to its own module
react-refresh/only-export-components lint error: exporting a non-component
function from a component file breaks Fast Refresh. Move the function to
mergePracticeNotesByTick.ts and import it in PracticeViewPlugin.tsx and
the test file.
